Change kernel modules from tar to squashfs and add it to boot media.
The dracut soci kernel module already had support for mounting kernel
modules at /krd/modules.squashfs to /lib/modules.  It will do that
if the file exists.

The big change here is in the bootkit produced layers, as now we
are creating a 'modules.squashfs' rather than 'modules.tar'.

In the future maybe we want to make oci-boot configurable to *not*
copy the modules, but right now we will just always copy the kernel
modules and the soci module mounts them if they are there.