Use KEYSET_D in substitution and keys-dir usage changes.
Note here we do:

    import:
      - path: ${{KEYSET_D}}/
        dest: /import/keys

My feeling is that this *should* produce /import/keys with the
same content as KEYSET_D.  However, the behavior of stacker 1.0.0-v4
is to create /import/keys/$(basename ${{KEYSET_D}}).
I think that is incorrect behavior as we would like to specify
the target "dest" independent of the source path.

If that were fixed, then we wouldn't need the 'keydir=$(echo ...)'
but rather just explicitly set it to 'keydir=/import/keys'.

https://github.com/project-stacker/stacker/issues/455

Also here is to use 'keydir' consistently (some places had keysdir).