Install pcr7 info as cpio
I have no idea why this is happening, but when we simply insert
pcr7data directory as is, it *is* in the initrd (as can be verified
by extracting kernel.efi), but the kernel's initrd unpacker does not
create it.  I've tried moving it to /etc/, still doesn't show up.  My
only guess, although I see no such limitation in init/initramfs.c, is
that the long directory filenames are a problem.

To work around this, create /etc/pcr7data.cpio inside the initrd.cpio,
and unpack it during soci-cmdline.sh.

With this, the livecd testcase in my mos PR works.

Signed-off-by: Serge Hallyn <shallyn@cisco.com>