Allow MOSCTL_BINARY and ZOT_BINARY to be read from the environment.
This allows user that invokes make to influence the mosctl
and zot binary via MOSCTL_BINARY and ZOT_BINARY variables respectively.

The make variables will take the values present in the environment
or on the make command line if provided.

So either:

    export MOSCTL_BINARY=/tmp/my-mosctl
    export ZOT_BINARY=http://...

or

    make MOSCTL_BINARY=... ZOT_BINARY=...

There is a workaround for stacker issue
https://github.com/project-stacker/stacker/issues/456