temporary fix: shim: use empty guid in certlist
I've been getting wrong pcr7 values.  Joy points out that we
should be using a SignatureList, not a SignatureData, when
writing the UKI certs into the shim.  The only meaningful
difference in the written out bytes is that the GUID should be
all 0s.  So just make them all 0s here, until we come up with
the proper fix using the go-efi libs.

Signed-off-by: Serge Hallyn <shallyn@cisco.com>