Update stacker go imports from v0.40.1 to v1.0.0-rc5.
My original intent of doing this was to be able to bump
the anuvu/squashfs import so that it would build with go 1.21.

But between v0.40.1 and v1.0.0-rc5 (current latest release)
of stacker, stacker dropped the c-go dependency on anuvu/squashfs.

So, now we *could* turn that go version back to 1.21 if we wanted,
but I don't see a large reason for doing so.

Signed-off-by: Scott Moser <smoser@brickies.net>