fix some 'imports' errors
The stackerfiles were using import: but looking for files
under /stacker/imports.  You have to use imports: for that.

Also, imports: with an import of 'dracut/' will take all
the contents of dracut/ and import them, rather than
recursively import the dracut directory.  But we were
expecting the dracut directory.  So drop the trailing
slash.

Signed-off-by: Serge Hallyn <serge@hallyn.com>